gma tgl: Pcode CDCLK_PREPARE_FOR_CHANGE is not mandatory
Pcode mailbox `CDCLK_PREPARE_FOR_CHANGE' command is not available till
BIOS Reset CPL and may fail when called too early.

This command is actually not mandatory to set the CDCLK sucessfully on
this platform.  This patch keeps the `CDCLK_PREPARE_FOR_CHANGE' command
but ignores return errors if any.

Change-Id: I0514e404a7b580d1e35af7b39c1bd77183f11058
Signed-off-by: Jeremy Compostella <jeremy.compostella@intel.com>
Signed-off-by: Anil Kumar <anil.kumar.k@intel.com>